+++ src/main/java/de/alpharogroup/lang/ClassExtensions.java
@@ -3,24 +3,20 @@
  *
  * Copyright (C) 2015 Asterios Raptis
  *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 package de.alpharogroup.lang;
 
@@ -348,8 +344,8 @@ public final class ClassExtensions
 	 * @throws URISyntaxException
 	 *             is thrown if a string could not be parsed as a URI reference.
 	 */
-	public static List<File> getDirectoriesFromResources(@NonNull String path, final boolean isPackage)
-		throws IOException, URISyntaxException
+	public static List<File> getDirectoriesFromResources(@NonNull String path,
+		final boolean isPackage) throws IOException, URISyntaxException
 	{
 		if (isPackage)
 		{
@@ -665,7 +661,8 @@ public final class ClassExtensions
 			if (url.getProtocol().equals("jar"))
 			{
 				InputStream resourceAsStream = ClassExtensions.getResourceAsStream(name, obj);
-				file = new File(System.getProperty("java.io.tmpdir"), obj.getClass().getSimpleName());
+				file = new File(System.getProperty("java.io.tmpdir"),
+					obj.getClass().getSimpleName());
 				Files.copy(resourceAsStream, file.toPath(), StandardCopyOption.REPLACE_EXISTING);
 			}
 			if (url.getProtocol().equals("file"))
@@ -687,7 +684,8 @@ public final class ClassExtensions
 	 *            The uri as String.
 	 * @return The InputStream from the uri.
 	 */
-	public static InputStream getResourceAsStream(final @NonNull Class<?> clazz, final @NonNull String uri)
+	public static InputStream getResourceAsStream(final @NonNull Class<?> clazz,
+		final @NonNull String uri)
 	{
 		InputStream is = clazz.getResourceAsStream(uri);
 		if (null == is)
@@ -720,7 +718,8 @@ public final class ClassExtensions
 	 *            The Object.
 	 * @return The resource or null if the resource does not exists.
 	 */
-	public static InputStream getResourceAsStream(final @NonNull String name, final @NonNull Object obj)
+	public static InputStream getResourceAsStream(final @NonNull String name,
+		final @NonNull Object obj)
 	{
 		InputStream inputStream = obj.getClass().getResourceAsStream(name);
 		if (null == inputStream)