+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 ## Change log
 ----------------------
 
+Version 3.2
+-------------
+
+CHANGED:
+
+- update of parent version to 5
+- update of silly-collections version to 5.2.1
+- update of test-objects dependency version to 5.2
+- update of cglib dependency version to 3.2.12
+
+Version 3.1
+-------------
+
+ADDED:
+ 
+- new method created for create new instance of an array with reflection with capacity
+- new method created for copy an existing array 
+- new methods created for set a field value from a given source object over a field and field name
+
 Version 3
 -------------